+++ web/src/utils/time.ts
@@ -1,0 +1,5 @@
+/** "HH:MM" → 0時からの経過分数 */
+export function timeToMinutes(time: string): number {
+  const [h, m] = time.split(':').map(Number);
+  return h * 60 + m;
+}
